# openapi.yaml
openapi: 3.2.0
info:
  title: Automater Factory API
  version: 0.1.0
  description: |
    SPA backend for blog automation campaign management.
    Designed from current ERD + codebase.
  license:
    name: Proprietary
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    pathId:
      name: id
      in: path
      required: true
      schema:
        type: integer
    pageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    perPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
  schemas:
...
    PostLayout:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        slots:
          type: array
          items:
            $ref: "#/components/schemas/LayoutSlot"
        created_at:
          type: string
          format: date-time
    LayoutSlot:
      type: object
      properties:
        id:
          type: integer
        layout_id:
          type: integer
        sort_order:
          type: integer
        block_type:
          type: string
          enum: [heading, paragraph, image, featured, list, quote, divider]
        config:
          type: ["object", "null"]
          description: Block fields with {placeholder} support
...
    RenderedBlock:
      type: object
      properties:
        block_type:
          type: string
        rendered_config:
          type: object
...
    Media:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        original_name:
          type: string
        content_type:
          type: string
        storage_path:
          type: string
        size_bytes:
          type: integer
        created_at:
          type: string
          format: date-time
...
paths:
...
  /layouts:
    get:
      tags: [layouts]
      operationId: listLayouts
      summary: My post layouts
      responses:
        "200":
          description: Layout list without slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PostLayout"
    post:
      tags: [layouts]
      operationId: createLayout
      summary: Create layout with optional initial slots
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
                slots:
                  type: array
                  items:
                    type: object
                    required: [sort_order, block_type]
                    properties:
                      sort_order:
                        type: integer
                      block_type:
                        type: string
                        enum: [heading, paragraph, image, featured, list, quote, divider]
                      config:
                        type: object
      responses:
        "201":
          description: Created with slots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostLayout"
  /layouts/{id}:
    get:
      tags: [layouts]
      operationId: getLayout
      summary: Layout with all slots
      parameters:
        - $ref: "#/components/parameters/pathId"
      responses:
        "200":
          description: Layout with slots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostLayout"
    patch:
      tags: [layouts]
      operationId: updateLayout
      parameters:
        - $ref: "#/components/parameters/pathId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostLayout"
    delete:
      tags: [layouts]
      operationId: deleteLayout
      summary: Delete layout (CASCADE deletes slots)
      parameters:
        - $ref: "#/components/parameters/pathId"
      responses:
        "204":
          description: Deleted
  /layouts/{id}/slots:
    put:
      tags: [layouts]
      operationId: replaceSlots
      summary: Replace all slots at once (drag-and-drop save)
      parameters:
        - $ref: "#/components/parameters/pathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                required: [sort_order, block_type]
                properties:
                  sort_order:
                    type: integer
                  block_type:
                    type: string
                    enum: [heading, paragraph, image, featured, list, quote, divider]
                  config:
                    type: object
      responses:
        "200":
          description: New slot list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LayoutSlot"
  /layouts/{id}/preview:
    post:
      tags: [layouts]
      operationId: previewLayout
      summary: Render layout with sample keywords
      parameters:
        - $ref: "#/components/parameters/pathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                values:
                  type: object
                  description: "slug to value map, e.g. {region: '강남', subject: '수학'}"
      responses:
        "200":
          description: Rendered block list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RenderedBlock"

  # ── Publish Presets ──────────────────────────────────────────────────────
...
  /media:
    get:
      tags: [media]
      operationId: listMedia
      summary: My uploaded media files
      parameters:
        - name: content_type
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/perPageParam"
      responses:
        "200":
          description: Media list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Media"
    post:
      tags: [media]
      operationId: uploadMedia
      summary: Upload an image file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Media record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Media"
  /media/{id}:
    get:
      tags: [media]
      operationId: getMedia
      summary: Get media metadata
      parameters:
        - $ref: "#/components/parameters/pathId"
      responses:
        "200":
          description: Media detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Media"
    delete:
      tags: [media]
      operationId: deleteMedia
      summary: Delete media record and file
      parameters:
        - $ref: "#/components/parameters/pathId"
      responses:
        "204":
          description: Deleted
  /media/{id}/file:
    get:
      tags: [media]
      operationId: serveMedia
      summary: Serve the actual file for download/display
      parameters:
        - $ref: "#/components/parameters/pathId"
      responses:
        "200":
          description: File content
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary